Documentation and support of string as user event definition
To use the string definition as user event type define the macro
USE_STRING_EVENT_TYPE in your compiler preprocessor (or in your prodject
property if using an IDE).
the application don't work properly if the macro is define in another
place than preprocesor macro because multiple class use the eventHandler
redefinition so it' complicated to change multiple file instead of one
prodject option.